requirements.txt
# requirements.txt
pillow
pycryptodome
numpy